+++ unixlauncher/launcher.py
@@ -84,7 +84,7 @@ class LoadingWidget(QtWidgets.QMainWindow, Ui_Downloader):
     def start_download(self):
         self.label.setText('Идёт загрузка игры')
         try:
-            response = requests.get(f'http://{self.host}:{self.port}/static/releases/unixgame.zip', stream=True)
+            response = requests.get(f'http://{self.host}:{self.port}/static/releases/game.zip', stream=True)
             chunks_required = round(int(response.headers.get('content-length', '0')) / 4096)
             if chunks_required:
                 chunks_downloaded = 0
@@ -131,26 +131,26 @@ class Ui_MainWindow(object):
         self.pushButton_2 = QtWidgets.QPushButton(self.centralwidget)
         self.pushButton_2.setGeometry(QtCore.QRect(420, 350, 181, 61))
         self.pushButton_2.setStyleSheet(":!hover{\n"
-                                        "background-color: rgb(172, 216, 230);\n"
-                                        "}\n"
-                                        "\n"
-                                        ":hover{\n"
-                                        "    border: 4px solid black; \n"
-                                        "    background: rgb(105, 185, 211); \n"
-                                        "    padding: 10px;\n"
-                                        "}")
+"background-color: rgb(172, 216, 230);\n"
+"}\n"
+"\n"
+":hover{\n"
+"    border: 4px solid black; \n"
+"    background: rgb(105, 185, 211); \n"
+"    padding: 10px;\n"
+"}")
         self.pushButton_2.setObjectName("pushButton_2")
         self.pushButton_3 = QtWidgets.QPushButton(self.centralwidget)
         self.pushButton_3.setGeometry(QtCore.QRect(520, 430, 171, 61))
         self.pushButton_3.setStyleSheet("backgr:!hover{\n"
-                                        "background-color: rgb(172, 216, 230);\n"
-                                        "}\n"
-                                        "\n"
-                                        ":hover{\n"
-                                        "    border: 4px solid black; \n"
-                                        "    background: rgb(105, 185, 211); \n"
-                                        "    padding: 10px;\n"
-                                        "}ound-color: rgb(172, 216, 230);")
+"background-color: rgb(172, 216, 230);\n"
+"}\n"
+"\n"
+":hover{\n"
+"    border: 4px solid black; \n"
+"    background: rgb(105, 185, 211); \n"
+"    padding: 10px;\n"
+"}ound-color: rgb(172, 216, 230);")
         self.pushButton_3.setObjectName("pushButton_3")
         self.plainTextEdit = QtWidgets.QPlainTextEdit(self.centralwidget)
         self.plainTextEdit.setGeometry(QtCore.QRect(10, 340, 301, 151))
@@ -158,28 +158,27 @@ class Ui_MainWindow(object):
         self.pushButton_5 = QtWidgets.QPushButton(self.centralwidget)
         self.pushButton_5.setGeometry(QtCore.QRect(320, 430, 191, 61))
         self.pushButton_5.setStyleSheet(":!hover{\n"
-                                        "background-color: rgb(172, 216, 230);\n"
-                                        "}\n"
-                                        "\n"
-                                        ":hover{\n"
-                                        "    border: 4px solid black; \n"
-                                        "    background: rgb(105, 185, 211); \n"
-                                        "    padding: 10px;\n"
-                                        "}")
+"background-color: rgb(172, 216, 230);\n"
+"}\n"
+"\n"
+":hover{\n"
+"    border: 4px solid black; \n"
+"    background: rgb(105, 185, 211); \n"
+"    padding: 10px;\n"
+"}")
         self.pushButton_5.setObjectName("pushButton_5")
         self.label_3 = QtWidgets.QLabel(self.centralwidget)
-        self.label_3.setGeometry(QtCore.QRect(10, 80, 281, 21))
+        self.label_3.setGeometry(QtCore.QRect(10, 80, 681, 21))
         font = QtGui.QFont()
         font.setPointSize(12)
         self.label_3.setFont(font)
         self.label_3.setObjectName("label_3")
         self.label_4 = QtWidgets.QLabel(self.centralwidget)
-        self.label_4.setGeometry(QtCore.QRect(10, 0, 681, 51))
+        self.label_4.setGeometry(QtCore.QRect(10, 10, 681, 61))
         font = QtGui.QFont()
         font.setPointSize(12)
         self.label_4.setFont(font)
         self.label_4.setStyleSheet("background-color: rgb(172, 216, 230);")
-        self.label_4.setTextFormat(QtCore.Qt.AutoText)
         self.label_4.setObjectName("label_4")
         self.tabWidget = QtWidgets.QTabWidget(self.centralwidget)
         self.tabWidget.setGeometry(QtCore.QRect(10, 110, 301, 221))
@@ -203,14 +202,14 @@ class Ui_MainWindow(object):
         self.pushButton = QtWidgets.QPushButton(self.tab)
         self.pushButton.setGeometry(QtCore.QRect(100, 130, 81, 51))
         self.pushButton.setStyleSheet(":!hover{\n"
-                                      "background-color: rgb(172, 216, 230);\n"
-                                      "}\n"
-                                      "\n"
-                                      ":hover{\n"
-                                      "    border: 4px solid black; \n"
-                                      "    background: rgb(105, 185, 211); \n"
-                                      "    padding: 10px;\n"
-                                      "}")
+"background-color: rgb(172, 216, 230);\n"
+"}\n"
+"\n"
+":hover{\n"
+"    border: 4px solid black; \n"
+"    background: rgb(105, 185, 211); \n"
+"    padding: 10px;\n"
+"}")
         self.pushButton.setObjectName("pushButton")
         self.tabWidget.addTab(self.tab, "")
         self.tab_2 = QtWidgets.QWidget()
@@ -239,14 +238,14 @@ class Ui_MainWindow(object):
         self.pushButton_4 = QtWidgets.QPushButton(self.tab_2)
         self.pushButton_4.setGeometry(QtCore.QRect(50, 130, 181, 51))
         self.pushButton_4.setStyleSheet(":!hover{\n"
-                                        "background-color: rgb(172, 216, 230);\n"
-                                        "}\n"
-                                        "\n"
-                                        ":hover{\n"
-                                        "    border: 4px solid black; \n"
-                                        "    background: rgb(105, 185, 211); \n"
-                                        "    padding: 10px;\n"
-                                        "}")
+"background-color: rgb(172, 216, 230);\n"
+"}\n"
+"\n"
+":hover{\n"
+"    border: 4px solid black; \n"
+"    background: rgb(105, 185, 211); \n"
+"    padding: 10px;\n"
+"}")
         self.pushButton_4.setObjectName("pushButton_4")
         self.tabWidget.addTab(self.tab_2, "")
         MainWindow.setCentralWidget(self.centralwidget)
@@ -270,7 +269,7 @@ class Ui_MainWindow(object):
         self.pushButton_2.setText(_translate("MainWindow", "Открыть сайт"))
         self.pushButton_3.setText(_translate("MainWindow", "Запустить игру!"))
         self.plainTextEdit.setPlainText(_translate("MainWindow", "Error logs:\n"
-                                                                 "OK"))
+"OK"))
         self.pushButton_5.setText(_translate("MainWindow", "Запустить мультиплеер!"))
         self.label_3.setText(_translate("MainWindow", "Auth status here"))
         self.label_4.setText(_translate("MainWindow", "Все системы функционируют нормально."))
@@ -286,6 +285,8 @@ class Ui_MainWindow(object):
         self.actionExit.setText(_translate("MainWindow", "Exit"))
 
 
+
+
 class MyWidget(QMainWindow, Ui_MainWindow):
     """ host = '130.193.46.251'
         port = '8080'
@@ -298,6 +299,9 @@ class MyWidget(QMainWindow, Ui_MainWindow):
         self.setupUi(self)
         self.initUI()
         self.update_game()
+        if not self.update_needed:
+            self.show()
+
 
     def initUI(self):
         self.pushButton_3.clicked.connect(self.launch_single)
@@ -312,7 +316,6 @@ class MyWidget(QMainWindow, Ui_MainWindow):
         self.user = None
         self.password = None
 
-        self.show()
 
     def check_hash(self, script_path=str(os.path.join(os.path.dirname(__file__), "COVIDcover"))):
         lst = []
@@ -328,22 +331,22 @@ class MyWidget(QMainWindow, Ui_MainWindow):
             response = requests.get(f'http://{MyWidget.host}:{MyWidget.port}/game_api/check_unix_hash/{output}')
         except requests.exceptions.ConnectionError:
             self.show_error('Отсутствует интернет. Запустите программу позже.')
-            return False
+            return True
         except requests.exceptions.Timeout:
             self.show_error('Видимо, у наш сервер сейчас отдыхает ;)')
-            return False
+            return True
         except Exception:
             self.show_error('Возникла непредвиденная ошибка. Вы можете написать в тех. поддержку.')
-            return False
+            return True
         if response.status_code == 500:
             self.show_error('Ошибка на сервере. Мы уже работаем.')
-            return False
+            return True
         if response.status_code != 200:
             self.show_error('Возникла непредвиденная ошибка. Вы можете написать в тех. поддержку.')
-            return False
+            return True
         if response:
             return response.json()['success']
-        return False
+        return True
 
     def register(self):
         email = self.lineEdit_3.text()
@@ -353,8 +356,7 @@ class MyWidget(QMainWindow, Ui_MainWindow):
             self.show_error('Заполните все поля!')
             return
         try:
-            response = requests.post(f'http://{MyWidget.host}:{MyWidget.port}/api/users/token/{user_registration}',
-                                     data={'email': email, 'password': password, "username": username})
+            response = requests.post(f'http://{MyWidget.host}:{MyWidget.port}/api/users/token/{user_registration}', data={'email': email, 'password': password, "username": username})
         except requests.exceptions.ConnectionError:
             self.show_error('Отсутствует интернет. Запустите программу позже.')
             return
@@ -383,6 +385,7 @@ class MyWidget(QMainWindow, Ui_MainWindow):
             return
         self.auth(email, password)
 
+
     def auth(self, email, password):
         try:
             response = requests.get(f'http://{MyWidget.host}:{MyWidget.port}/game_api/auth',
@@ -409,7 +412,7 @@ class MyWidget(QMainWindow, Ui_MainWindow):
         return True
 
     def update_game(self):
-        self.update_needed = True
+        self.update_needed = False
         if os.path.isfile(os.path.join(os.path.dirname(__file__), 'COVIDcover/versions.json')):
             try:
                 response = requests.get(f'http://{MyWidget.host}:{MyWidget.port}/static/releases/versions.json',
@@ -421,24 +424,22 @@ class MyWidget(QMainWindow, Ui_MainWindow):
                 self.show_error('Видимо, у наш сервер сейчас отдыхает ;) Вам недоступна сетевая игра')
                 return
             except Exception:
-                self.show_error(
-                    'Возникла непредвиденная ошибка.\nВы можете написать в тех. поддержку.\nВам недоступна сетевая игра')
+                self.show_error('Возникла непредвиденная ошибка.\nВы можете написать в тех. поддержку.\nВам недоступна сетевая игра')
                 return
             if response.status_code == 500:
                 self.show_error('Ошибка на сервере. Мы уже работаем.\nВам недоступна сетевая игра')
                 return False
             versions_list = json.loads(response.content.decode('utf-8'))
-            with open(os.path.join(os.path.dirname(__file__), 'COVIDcover/versions.json'), mode='r',
-                      encoding='utf-8') as f:
+            with open(os.path.join(os.path.dirname(__file__), 'COVIDcover/versions.json'), mode='r', encoding='utf-8') as f:
                 data = json.load(f)
                 self.update_needed = data['last_version'] != versions_list['last_version']
-
         if not self.update_needed:
             self.update_needed = not self.check_hash()
         if self.update_needed:
             self.hide()
-            self.updater = LoadingWidget(self)  # TODO Update launcher
+            self.updater = LoadingWidget(self) # TODO Update launcher
             self.updater.closed_signal.connect(self.show)
+
 
     def launch_multi(self):
         self.show_error('Все системы функционируют нормально.')
@@ -477,13 +478,18 @@ class MyWidget(QMainWindow, Ui_MainWindow):
             return
 
         try:
+            error_code = None
             self.hide()
-            os.system(f"{sys.executable} \"{os.path.join(os.path.abspath(os.path.dirname(__file__)), 'COVIDcover', 'multi_build/multi_main.py')}\" {data['ip']} {data['port']} {self.user['token']} {self.user['username']}")
+            os.system(f"cd COVIDcover && {sys.executable} \"multi_build/multi_main.exe\" {data['ip']} {data['port']} {self.user['token']} {self.user['username']}")
             self.show()
-            with open('score.dat', mode='r', encoding='utf-8') as file:
-                score, error_code = file.read().strip().split()
-                error_code = int(error_code)
-            os.remove('score.dat')
+            if os.path.isfile('score.dat'):
+                with open('score.dat', mode='r', encoding='utf-8') as file:
+                    score, error_code = file.read().strip().split()
+                    error_code = int(error_code)
+                os.remove('score.dat')
+            else:
+                score, error_code = 0, -7
+
             if not score.isdigit():
                 self.show_error('Ошибка клиента игры. Напишите нам.')
             if error_code == -5:
@@ -510,7 +516,9 @@ class MyWidget(QMainWindow, Ui_MainWindow):
             finally:
                 if os.path.isfile('score.dat'):
                     os.remove('score.dat')
-                if error_code != 0:
+                if os.path.isfile('covid_cover.log'):
+                    os.remove('covid_cover.log')
+                if error_code and error_code != 0:
                     if os.path.isfile(log_filename):
                         try:
                             os.remove(log_filename)
@@ -519,11 +527,9 @@ class MyWidget(QMainWindow, Ui_MainWindow):
 
                     with open(log_filename, mode='r', encoding='utf-8') as file:
                         try:
-                            response = requests.post(f'http://{MyWidget.host}:{MyWidget.port}/game_api/get_log',
-                                                     files={'log': file})
+                            response = requests.post(f'http://{host}:{port}/game_api/get_log', files={'log': file})
                         except Exception as e:
                             self.show_error(str(e))
-                    pass
                 if os.path.isfile(log_filename):
                     try:
                         os.remove(log_filename)
@@ -535,7 +541,7 @@ class MyWidget(QMainWindow, Ui_MainWindow):
     def launch_single(self):
         try:
             self.hide()
-            os.system(f"{sys.executable} \"{os.path.join(os.path.abspath(os.path.dirname(__file__)), 'COVIDcover', 'main_build/main.py')}\"")
+            os.system(f'cd COVIDcover && {sys.executable} "main_build/main"')
             self.show()
         except Exception as e:
             self.plainTextEdit.setPlainText(self.plainTextEdit.toPlainText() + str(e) + '\n\n')
